requirements.txt
Flask==3.0.0
scapy==2.5.0
geoip2==4.8.0
PyYAML==6.0.1
requests==2.31.0
python-dotenv==1.0.0
elasticsearch==8.15.0
Werkzeug==3.0.1
